Fixes transfers for MKV conformance check tests
The modified transfers in when-normalized-none-valid/ and
when-normalized-all-valid/ should now conform to their descriptive names: the
MKV generated (via ffmpeg) from the file in the former should be invalid and
that generated from the file in the latter should be valid (according to
MediaConch). For some reason, ffmpeg is now generating an invalid MKV from
mp4-test.mp4 even though this behaviour was not observed previously.
Because AM MediaConch is still at 16.12, I surmise that an update in
ffmpeg is responsible for this change.